Renamed utils/solid to utils/network
